--- src/widgets/header/header-desktop/index.ts
@@ -1,2 +1,0 @@
-export * from './header-desktop'
-export * from './notifications-dropdown'
