+++ src/widgets/header/index.ts
@@ -1,3 +1,2 @@
 export * from './header'
-export * from './header-desktop'
-export * from './header-mobile'
+export * from './ui'
