+++ src/widgets/header/ui/auth-buttons/index.ts
@@ -1,0 +1,1 @@
+export * from './auth-buttons'
