+++ src/widgets/header/ui/header-desktop/index.ts
@@ -1,0 +1,2 @@
+export * from './header-desktop'
+export * from '../notifications-dropdown'
