+++ src/widgets/header/ui/index.ts
@@ -1,0 +1,7 @@
+export * from './logo'
+export * from './header-mobile'
+export * from './header-desktop'
+export * from './auth-buttons'
+export * from './notifications-dropdown'
+export * from './notifications-dropdown-list'
+export * from './mobile-dropdown'
