+++ src/widgets/header/ui/logo/index.ts
@@ -1,0 +1,1 @@
+export * from './logo'
