+++ src/widgets/header/ui/mobile-dropdown/index.ts
@@ -1,0 +1,1 @@
+export * from './mobile-dropdown'
