+++ src/widgets/header/ui/notifications-dropdown-list/index.ts
@@ -1,0 +1,1 @@
+export * from './notifications-dropdown-list'
