+++ src/widgets/header/ui/notifications-dropdown/index.ts
